// src/app/models/assigners.ts
export class Assigners {
    id: number;
    userid: number;
    name: string;
    username: string;
    email: string;
    testid: number;
    testname: String;
    testdate: Date;
    questions1: string;
    questions2: string;
    questions3: string;
    questions4: string;
    questions5: string;
    questions6: string;
    questions7: string;
    questions8: string;
    questions9: string;
    questions10: string;
    option1: any;
    option2: any;
    option3: any;
    option4: any;
  userregs: import('c:/Users/Sivakumar/Documents/MCQ/src/app/models/userregs').Userregs;
  questions: import('c:/Users/Sivakumar/Documents/MCQ/src/app/services/questions').Questions[];
}